Split E2E tests into 3 parallel shards with per-shard scripts
