+++ package.json
@@ -14,7 +14,11 @@
     "dev": "COSTGOBLIN_DATA_DIR=$PWD/data/processed COSTGOBLIN_CONFIG_DIR=$PWD/data/config npm run dev --workspace=packages/desktop",
     "dev:fixtures": "COSTGOBLIN_DATA_DIR=$PWD/packages/core/src/__fixtures__/synthetic COSTGOBLIN_CONFIG_DIR=$PWD/packages/core/src/__fixtures__/config npm run dev --workspace=packages/desktop",
     "build": "npm run build --workspaces",
-    "e2e": "npm run build --workspace=packages/desktop && npx playwright test && npx tsx e2e/collect-coverage.ts",
+    "e2e": "npm run build --workspace=packages/desktop && npx playwright test e2e/views-core.test.ts e2e/views-config.test.ts e2e/stress.test.ts && npx tsx e2e/collect-coverage.ts",
+    "e2e:core": "npm run build --workspace=packages/desktop && npx playwright test e2e/views-core.test.ts && npx tsx e2e/collect-coverage.ts",
+    "e2e:config": "npm run build --workspace=packages/desktop && npx playwright test e2e/views-config.test.ts && npx tsx e2e/collect-coverage.ts",
+    "e2e:stress": "npm run build --workspace=packages/desktop && npx playwright test e2e/stress.test.ts",
+    "e2e:all": "npm run build --workspace=packages/desktop && npx playwright test e2e/views-core.test.ts e2e/views-config.test.ts e2e/stress.test.ts && npx tsx e2e/collect-coverage.ts",
     "perf": "npm run build --workspace=packages/desktop && npx playwright test e2e/perf.test.ts",
     "prepare": "husky"
   },
